Tweaking test.sh

+++ test.sh
@@ -8,8 +8,10 @@ set -e
 ./validate.sh
 mkdir -p test-reports
 tox -e ${versions} -- ${posargs}
-coverage html --directory test-reports/coverage
-rm -rf test-reports/.coverage* test-reports/Running 2> /dev/null
+if [ $versions = "ALL" ]; then
+    coverage html --directory test-reports/coverage
+    rm -rf test-reports/.coverage* test-reports/Running 2> /dev/null
+fi
 
 duration=$SECONDS
 echo "$(($duration / 60)) minutes and $(($duration % 60)) seconds elapsed."
